requirements.txt
elasticsearch>=7.0.0
PyYAML 